Feat: Outsourcing context & state into separate provider component
- In this we have separated all the functions and useState from the app component to cart context
- Then we just wrapped that context function inside app component,
